refactor(runtime): give each operation one home on its handle
Every operation existed twice: a body on `FsCore` and a forwarder on the
handle that exposed it, each carrying its own copy of the doc comment. The
copies had already drifted -- `delete_path` told callers that physical
reclamation is "explicit garbage collection" on one side and "background
garbage collection" on the other.

Move each body onto the handle that owns it: reads onto `FsReader`,
maintenance onto `FsAdmin`, mutations and uploads onto `FsWriter`. One
method, one doc comment, one place.

The core layer stays, because two internal callers genuinely need it: the
publication service calls back into the write path, and writer-scheduled
maintenance runs a step. Both now go through a handle built with `from_core`,
mirroring the constructor `FsReader` already had -- so background maintenance
runs the operator's own step and a publish task runs the caller's own mutation
path, rather than private copies of them.

Resolving the drifted comment: "background" was the wrong survivor. Nothing
sweeps unless a caller opts in -- `MaintenanceStepOptions::default()` leaves
`gc` unset, and STYLE.md makes garbage collection an explicit `FsAdmin` call.
