+++ crates/loonfs/src/fs/uploads.rs
@@ -1,8 +1,8 @@
 //! Staged uploads and direct-put targets.
 
-use super::core::FsCore;
 use crate::publish::PreparedContent;
 use crate::uploads::BeginDirectPutUploadTargetResponse;
+use crate::FsWriter;
 use crate::Result;
 use crate::{
     BeginUploadRequest, BeginUploadResponse, CompleteUploadRequest, CompleteUploadResponse,
@@ -10,46 +10,49 @@ use crate::{
 };
 use loonfs_api::UploadId;
 
-impl FsCore {
+impl FsWriter {
     /// Starts a durable upload session for a namespace.
-    pub(crate) async fn begin_upload(
+    pub async fn begin_upload(
         &self,
         namespace_id: &NamespaceId,
         request: BeginUploadRequest,
     ) -> Result<BeginUploadResponse> {
         Ok(self
+            .core
             .namespace_engine(namespace_id)
             .begin_upload(request)
             .await?)
     }
 
     /// Starts a direct-put upload session and returns the internal target for server-side signing.
-    pub(crate) async fn begin_direct_put_upload_target(
+    pub async fn begin_direct_put_upload_target(
         &self,
         namespace_id: &NamespaceId,
         content_ref: ContentRef,
     ) -> Result<BeginDirectPutUploadTargetResponse> {
         Ok(self
+            .core
             .namespace_engine(namespace_id)
             .begin_direct_put_upload_target(content_ref)
             .await?)
     }
 
     /// Uploads whole-file content into an upload session.
-    pub(crate) async fn upload_content(
+    pub async fn upload_content(
         &self,
         namespace_id: &NamespaceId,
         upload_id: &UploadId,
         bytes: &[u8],
     ) -> Result<UploadContentResponse> {
         Ok(self
+            .core
             .namespace_engine(namespace_id)
             .upload_content(upload_id, bytes)
             .await?)
     }
 
     /// Completes an upload session when the expected content ref matches.
-    pub(crate) async fn complete_upload(
+    pub async fn complete_upload(
         &self,
         namespace_id: &NamespaceId,
         upload_id: &UploadId,
@@ -65,7 +68,7 @@ impl FsCore {
     ///
     /// Service-proxied completion performs no content-blob I/O. Direct-put
     /// completion performs one content-blob HEAD and no content-blob GET.
-    pub(crate) async fn complete_upload_prepared(
+    pub async fn complete_upload_prepared(
         &self,
         namespace_id: &NamespaceId,
         upload_id: &UploadId,
@@ -75,6 +78,7 @@ impl FsCore {
             .load_namespace_catalog_for_content_preparation(namespace_id)
             .await?;
         Ok(self
+            .core
             .namespace_engine(namespace_id)
             .complete_upload_prepared_with_catalog(&catalog, upload_id, request)
             .await?)
